apps/pyproject.toml
[project]
name = "edugraph"
version = "0.1.0"
description = "EduGraph App"
authors = [{ name = "Christian Bick", email = "christian.bick@bitsuppliers.com" }]
requires-python = "~=3.12"
dependencies = [
    "owlready2>=0.46,<0.47",
    "flask~=3.0",
    "python-dotenv>=1.0.1,<2",
    "flask-cors>=5.0.0,<6",
    "waitress>=3.0.1,<4",
    "google-genai>=1.19.0,<2",
    "google-cloud-storage>=1.44.0,<2",
    "numpy>=2.3.0",
    "google-cloud-aiplatform>=1.96.0",
    "pymupdf>=1.26.0",
    "setuptools>=80.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "sementic"]

[dependency-groups]
dev = [
    "pytest~=8.3",
    "assertpy~=1.1",
]

